+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.3.0]
+
+### Changed
+
+- Update to Bevy v0.13
+
+
 ## [0.2.0]
 
 ### Added